+++ pdf/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Add support for table spans [gustl22]
+
 ## 3.11.4
 
 - Improve performance of table layout [Pieter van Loon]